refactor: set up scaffolding of routing
