+++ index.html
@@ -33,7 +33,9 @@
     <meta property="twitter:image" content="https://control-records.netlify.app/preview.png">
   </head>
   <body>
-    <div id="app"></div>
+    <div id="app">
+      <router-view></router-view>
+    </div>
     <script type="module" src="/src/main.js"></script>
   </body>
 </html>
