+++ src/components/HeaderBlock.vue
@@ -1,0 +1,17 @@
+<template>
+  <header>
+    <h1>Federal Bureau of Control / Records Archive System</h1>
+  </header>
+</template>
+
+<style scoped>
+h1 {
+  margin-top: 0;
+}
+
+@media only print {
+  header {
+    display: none;
+  }
+}
+</style>
